workflows/pyproject.toml
[project]
name = "workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.5",
    "google-adk>=1.28.1",
    "numpy>=2.4.4",
    "reportlab>=4.4.10",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "tenacity>=9.1.4",
]
